+++ src/sorting/bubble_sort.hpp
@@ -4,4 +4,4 @@
 using namespace std;
 
 void Bubble_Sort(int* arr, int n);
-void Bubble_Sort_Count(int* arr, int n, int& count);
+void Bubble_Sort_Count(int* arr, int n, long long& count);
