+++ src/sorting/counting_sort.hpp
@@ -4,4 +4,4 @@
 using namespace std;
 
 void Counting_Sort(int* arr, int n);
-void Counting_Sort_Count(int* arr, int n, int& count);
+void Counting_Sort_Count(int* arr, int n, long long& count);
