+++ src/sorting/selection_sort.hpp
@@ -4,4 +4,4 @@
 using namespace std;
 
 void Selection_Sort(int* arr, int n);
-void Selection_Sort_Count(int* arr, int n, int& count);
+void Selection_Sort_Count(int* arr, int n, long long& count);
