+++ src/sorting/shaker_sort.hpp
@@ -4,4 +4,4 @@
 using namespace std;
 
 void Shaker_Sort(int* arr, int n);
-void Shaker_Sort_Count(int* arr, int n, int& count);
+void Shaker_Sort_Count(int* arr, int n, long long& count);
