+++ src/sorting/shell_sort.hpp
@@ -4,4 +4,4 @@
 using namespace std;
 
 void Shell_Sort(int* arr, int n);
-void Shell_Sort_Count(int* arr, int n, int& count);
+void Shell_Sort_Count(int* arr, int n, long long& count);
